diploma-frontend/diploma_shop/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.4
django-filter==25.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
pillow==11.3.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
setuptools==80.9.0
simplejson==3.20.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
